use shared superclass
